requirements.txt
django<1.6
south
sorl-thumbnail
PIL
